Refactor name from Duke to Shima Create more classes in design package Move members from design.Default to the newly created classes Further improve code readability in main
